requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
CairoSVG==2.2.1
cffi==1.11.4
colorama==0.3.9
configparser==3.5.0
cryptography==2.1.4
diff-match-patch==20121119
Django==2.0.2
django-import-export==0.7.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-mptt-admin==0.5.0
django-registration==2.4.1
et-xmlfile==1.0.1
Fabric3==1.13.1.post1
gaffer==0.5.2
gunicorn==19.7.1
idna==2.6
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.0
paramiko==2.4.0
psutil==5.4.3
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2017.3
pyuv==1.4.0
PyYAML==3.12
Pillow==5.3.0
setproctitle==1.1.10
six==1.11.0
tablib==0.12.1
tornado==2.4.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
